Refactor: split utility functions
